// apps/web/src/api/generated/orval/model/createPartNumberCategoryRequest.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 채번 카테고리 생성 요청
 */
export interface CreatePartNumberCategoryRequest {
  /**
   * 카테고리 이름
   * @minLength 0
   * @maxLength 100
   */
  name: string;
  /**
   * 채번 접두어
   * @minLength 0
   * @maxLength 20
   */
  prefix: string;
  /**
   * 구분자
   * @minLength 0
   * @maxLength 5
   */
  delimiter?: string;
  /**
   * 자릿수
   * @minimum 1
   * @maximum 10
   */
  digits?: number;
}