// apps/web/src/api/generated/orval/model/updatePartNumberCategoryRequest.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 채번 카테고리 수정 요청
 */
export interface UpdatePartNumberCategoryRequest {
  /**
   * 카테고리 이름
   * @minLength 0
   * @maxLength 100
   */
  name: string;
  /**
   * 채번 접두어
   * @minLength 0
   * @maxLength 20
   */
  prefix: string;
  /**
   * 구분자
   * @minLength 0
   * @maxLength 5
   */
  delimiter?: string;
  /**
   * 자릿수
   * @minimum 1
   * @maximum 10
   */
  digits?: number;
}