// apps/web/src/api/generated/orval/model/partNumberCategoryListResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PartNumberCategoryResponse } from './partNumberCategoryResponse';

/**
 * 채번 카테고리 목록 응답
 */
export interface PartNumberCategoryListResponse {
  /** 채번 카테고리 목록 */
  items?: PartNumberCategoryResponse[];
}