// apps/web/src/api/generated/orval/model/partNumberCategoryResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 채번 카테고리 응답
 */
export interface PartNumberCategoryResponse {
  /** 채번 카테고리 ID */
  id?: string;
  /** 카테고리 이름 */
  name?: string;
  /** 채번 접두어 */
  prefix?: string;
  /** 구분자 */
  delimiter?: string;
  /** 자릿수 */
  digits?: number;
  /** 예시 품번 */
  preview_part_number?: string;
}