// apps/web/src/api/generated/orval/model/partNumberAvailabilityResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 품번 사용 가능 여부 응답
 */
export interface PartNumberAvailabilityResponse {
  /** 조회한 품번 */
  part_number?: string;
  /** 사용 가능 여부 */
  available?: boolean;
}